:bug: BugFix: 평균값 새로고침시 초기화 되는 현상
Related to #22
