+++ src/components/detail/ReviewTab.js
@@ -1,11 +1,10 @@
 import { faStar } from "@fortawesome/free-regular-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { styled } from "styled-components";
 import Modal from "react-modal";
 import ReviewModal from "./ReviewModal";
-import { collection, getDocs } from "firebase/firestore";
-import { db } from "../../firebase";
+import fetchUserReviewsRatings from "../utils/fetchUserReviewsRatings";
 
 // Modal react-modal doc
 Modal.setAppElement("#root");
@@ -22,31 +21,18 @@ const ReviewTab = () => {
     setIsopen(false);
   };
 
-  //firebase 데이터 Read
-  const fetchUserReviewsRatings = async () => {
-    try {
-      const userReviewsCollection = collection(db, "userReviews");
-      const querySnapshot = await getDocs(userReviewsCollection);
+  useEffect(() => {
+    const initializeAverageGrade = async () => {
+      try {
+        const average = await fetchUserReviewsRatings();
+        setAverageGrade(average);
+      } catch (error) {
+        console.error("Error initializing averageGrade: ", error);
+      }
+    };
 
-      const ratings = [];
-      querySnapshot.forEach((doc) => {
-        const data = doc.data();
-        const rating = data.grade;
-
-        // 별점이 있는 경우에만 배열에 추가
-        if (rating) {
-          ratings.push(rating);
-        }
-      });
-      const sum =
-        ratings.length > 0 ? ratings.reduce((acc, grade) => acc + grade, 0) : 0;
-      const average = ratings.length > 0 ? sum / ratings.length : 0;
-      return average;
-    } catch (error) {
-      console.error("Error fetching user reviews: ", error);
-      return [];
-    }
-  };
+    initializeAverageGrade();
+  }, []);
 
   return (
     <ReviewContainer>
